+++ config/application.rb
@@ -31,6 +31,7 @@ module AtuservicioRails
       visualization_components/specialities_visualization_cell
       visualization_components/in_hundreed_visualization_cell
       visualization_components/people_visualization_cell
+      components/provider_list_large_cell
       components/reference_map_cell
       components/site_header_cell
       components/story_slider_cell